+++ Elements/src/Geometry/Profiles/MCProfileFactory.cs
@@ -3,11 +3,11 @@ namespace Elements.Geometry.Profiles
     /// <summary>
     /// A factory for creation L profiles.
     /// </summary>
-    public class CProfileFactory : ParametricProfileFactory<CProfileType, CProfile>
+    public class MCProfileFactory : ParametricProfileFactory<MCProfileType, MCProfile>
     {
         /// <summary>
         /// Create an C profile factory.
         /// </summary>
-        public CProfileFactory() : base("./ProfileData/C.csv", Units.InchesToMeters(1)) { }
+        public MCProfileFactory() : base("./ProfileData/MC.csv", Units.InchesToMeters(1)) { }
     }
 }