+++ Elements/test/ParametricProfileTests.cs
@@ -160,61 +160,19 @@ namespace Elements.Geometry.Tests
             Model.AddElement(beam);
         }
 
-        [Fact]
-        public void LProfileFactoryCreate()
+        [Theory]
+        [InlineData("C", typeof(CProfileFactory))]
+        [InlineData("WT", typeof(WTProfileFactory))]
+        [InlineData("L", typeof(WTProfileFactory))]
+        [InlineData("ST", typeof(WTProfileFactory))]
+        [InlineData("MC", typeof(MCProfileFactory))]
+        public void ProfileFactory(string name, Type factoryType)
         {
-            Name = nameof(LProfileFactoryCreate);
+            Name = name;
 
-            var factory = new LProfileFactory();
-            var profiles = factory.AllProfiles();
-
-            var x = 0.0;
-            var z = 0.0;
-            foreach (var profile in profiles)
-            {
-                var line = new Line(new Vector3(x, 0, z), new Vector3(x, 3, z));
-                var beam = new Beam(line, profile);
-                Model.AddElement(beam);
-                x += 1.0;
-                if (x > 10.0)
-                {
-                    z += 1.0;
-                    x = 0.0;
-                }
-            }
-        }
-
-        [Fact]
-        public void WTProfileFactoryCreate()
-        {
-            Name = nameof(WTProfileFactoryCreate);
-
-            var factory = new WTProfileFactory();
-            var profiles = factory.AllProfiles();
-
-            var x = 0.0;
-            var z = 0.0;
-            foreach (var profile in profiles)
-            {
-                var line = new Line(new Vector3(x, 0, z), new Vector3(x, 3, z));
-                var beam = new Beam(line, profile);
-                Model.AddElement(beam);
-                x += 1.0;
-                if (x > 10.0)
-                {
-                    z += 1.0;
-                    x = 0.0;
-                }
-            }
-        }
-
-        [Fact]
-        public void CProfileFactoryCreate()
-        {
-            Name = nameof(CProfileFactoryCreate);
-
-            var factory = new CProfileFactory();
-            var profiles = factory.AllProfiles();
+            var factory = Activator.CreateInstance(factoryType);
+            var method = factoryType.GetMethod("AllProfiles");
+            var profiles = (IEnumerable<Profile>)method.Invoke(factory, null);
 
             var x = 0.0;
             var z = 0.0;